// src/redux/user/types.ts
export enum ActionType {
  LOGIN = 'USER/LOGIN',
  LOGIN_SUCCESS = 'USER/LOGIN_SUCCESS',
  LOGIN_ERROR = 'USER/LOGIN_ERROR',
  LOGOUT = 'USER/LOGOUT',
  LOGOUT_SUCCESS = 'USER/LOGOUT_SUCCESS',
  RESET_LOGIN_STATE = 'USER/RESET_LOGIN_STATE',
  INIT = 'USER/INIT',
  INIT_SUCCESS = 'USER/INIT_SUCCESS',
  SET_USER_INFO = 'USER/SET_USER_INFO',
  USER_CERTIFICATE = 'USER/USER_CERTIFICATE', // 实名认证
  USER_CERTIFICATE_SUCCESS = 'USER/USER_CERTIFICATE_SUCCESS',
  GET_WALLET_INFO = 'USER/GET_WALLET_INFO',
  GET_WALLET_INFO_SUCCESS = 'USER/GET_WALLET_INFO_SUCCESS',
  GET_SUPPORT_BANK_LIST = 'USER/GET_SUPPORT_BANK_LIST',
  LOAD_USER_INFO = 'USER/LOAD_USER_INFO',
  GET_SUPPORT_BANK_LIST_SUCCESS = 'USER/GET_SUPPORT_BANK_LIST_SUCCESS',
}